Add an arg to treat too-early date in HTTP response as error.
Some buggy routers changes 'Date:' header based on their own timers.
When a device is connected to such a router, and we reboot them at the same time,
htpdate on that device MAY receive a 1970-year date before the router getting an
NTP response, and cannot change to correct date later due to the huge offset.